Segregated sell token logic into dedicated hook
