+++ packages/commonwealth/client/scripts/views/modals/TradeTokenModel/TradeTokenForm/useTradeTokenForm.ts
@@ -1,20 +1,15 @@
 import { ExtendedCommunity } from '@hicommonwealth/schemas';
 import { commonProtocol } from '@hicommonwealth/shared';
-import { notifyError, notifySuccess } from 'controllers/app/notifications';
 import useRunOnceOnCondition from 'hooks/useRunOnceOnCondition';
 import NodeInfo from 'models/NodeInfo';
 import { useMemo, useState } from 'react';
 import { useGetCommunityByIdQuery } from 'state/api/communities';
-import useSellTokenMutation from 'state/api/launchPad/sellToken';
 import { fetchCachedNodes } from 'state/api/nodes';
-import {
-  useCreateTokenTradeMutation,
-  useGetERC20BalanceQuery,
-} from 'state/api/tokens';
 import useUserStore from 'state/ui/user';
 import { z } from 'zod';
 import { TradingMode, UseTradeTokenFormProps } from './types';
 import useBuyTrade from './useBuyTrade';
+import useSellTrade from './useSellTrade';
 
 const COMMON_PLATFORM_FEE_PERCENTAGE = 5; // make configurable when needed
 
@@ -58,9 +53,6 @@ const useTradeTokenForm = ({
     shouldRun: userAddresses.length > 0 && !selectedAddress,
   });
 
-  const { mutateAsync: createTokenTrade, isLoading: isCreatingTokenTrade } =
-    useCreateTokenTradeMutation();
-
   const onTradingModeChange = (mode: TradingMode) => {
     setTradingMode(mode);
   };
@@ -83,80 +75,23 @@ const useTradeTokenForm = ({
     tokenCommunity: tokenCommunity as z.infer<typeof ExtendedCommunity>,
   });
 
-  // sell mode logic start --- {
-  const [tokenSellAmount, setTokenSellAmount] = useState<number>(0); // can be fractional
-
-  const { mutateAsync: sellToken, isLoading: isSellingToken } =
-    useSellTokenMutation();
-
-  const handleTokenSell = async () => {
-    try {
-      // this condition wouldn't be called, but adding to avoid typescript issues
-      if (
-        !baseNode?.url ||
-        !baseNode?.ethChainId ||
-        !selectedAddress ||
-        !tokenCommunity
-      ) {
-        return;
-      }
-
-      // buy token on chain
-      const payload = {
-        chainRpc: baseNode.url,
-        ethChainId: baseNode.ethChainId,
-        amountToken: tokenSellAmount * 1e18, // amount in wei // TODO
-        walletAddress: selectedAddress,
-        tokenAddress: tradeConfig.token.token_address,
-      };
-      const txReceipt = await sellToken(payload);
-
-      // create token trade on db
-      await createTokenTrade({
-        eth_chain_id: baseNode?.ethChainId,
-        transaction_hash: txReceipt.transactionHash,
-      });
-
-      // update user about success
-      notifySuccess('Transactions successful!');
-
-      onTradeComplete?.();
-    } catch (e) {
-      notifyError('Failed to sell token');
-      console.log('Failed to sell token => ', e);
-    }
-  };
-
   const {
-    data: selectedAddressTokenBalance = `0.0`,
-    isLoading: isLoadingUserTokenBalance,
-  } = useGetERC20BalanceQuery({
-    nodeRpc: tokenCommunity?.ChainNode?.url || '',
-    tokenAddress: tradeConfig.token.token_address,
-    userAddress: selectedAddress || '',
+    amounts: sellTradeAmounts,
+    handleTokenSell,
+    isSellActionPending,
+    selectedAddressTokenBalance,
+  } = useSellTrade({
+    chainNode: baseNode,
+    selectedAddress,
+    commonFeePercentage: COMMON_PLATFORM_FEE_PERCENTAGE,
+    tradeConfig,
+    onTradeComplete,
+    tokenCommunity: tokenCommunity as z.infer<typeof ExtendedCommunity>,
   });
-
-  const onTokenSellAmountChange = (
-    change: React.ChangeEvent<HTMLInputElement>,
-  ) => {
-    const value = change.target.value;
-
-    if (value === '')
-      setTokenSellAmount(0); // TODO: fix decimal
-    // verify only numbers with decimal (optional) are present
-    else if (/^\d*(\.\d+)?$/.test(value)) {
-      setTokenSellAmount(parseFloat(value));
-    }
-  };
-  // sell mode logic end --- }
 
   // flag to indicate if something is ongoing
   const isActionPending =
-    isLoadingTokenCommunity ||
-    isBuyActionPending ||
-    isLoadingUserTokenBalance ||
-    isSellingToken ||
-    isCreatingTokenTrade;
+    isLoadingTokenCommunity || isBuyActionPending || isSellActionPending;
 
   const onCTAClick = () => {
     if (isActionPending) return;
@@ -175,31 +110,12 @@ const useTradeTokenForm = ({
   };
 
   return {
-    // Note: not exporting state setters directly, since some extra
-    // functionality is done in most "onChange" handlers above
+    // Note: not exporting state setters directly, all "buy/sell token"
+    // business logic should be done (or exported from) in this hook
     trading: {
       amounts: {
         buy: buyTradeAmounts,
-        sell: {
-          invest: {
-            // not to be confused with "Base" network on ethereum
-            baseToken: {
-              amount: tokenSellAmount,
-              onAmountChange: onTokenSellAmountChange,
-              unitEthExchangeRate: 100, // TODO: hardcoded for now - blocked token pricing
-              toEth: 100, // TODO: hardcoded for now - blocked token pricing
-            },
-            insufficientFunds:
-              tokenSellAmount > parseFloat(selectedAddressTokenBalance),
-            commonPlatformFee: {
-              percentage: `${COMMON_PLATFORM_FEE_PERCENTAGE}%`,
-              eth: 100, // TODO: hardcoded for now - blocked token pricing
-            },
-          },
-          gain: {
-            eth: 100, // TODO: hardcoded for now - blocked token pricing
-          },
-        },
+        sell: sellTradeAmounts,
       },
       mode: { value: tradingMode, onChange: onTradingModeChange },
       token: tradeConfig.token,
@@ -216,8 +132,8 @@ const useTradeTokenForm = ({
             isLoading: selectedAddressEthBalance.isLoading,
           },
           selectedToken: {
-            value: selectedAddressTokenBalance,
-            isLoading: isLoadingUserTokenBalance,
+            value: selectedAddressTokenBalance.value,
+            isLoading: selectedAddressTokenBalance.isLoading,
           },
         },
         onChange: onChangeSelectedAddress,
